made subcomponent for question information
